app/requirements.txt
streamlit
transformers
torch
openai
numpy
pandas
diffusers
torch
Pillow